- Fix a problem with ksystraycmd using window frame instead of the contents   in Wid mode. Strictly speaking, the old behaviour was possibly correct,   but it certainly isn't the behaviour you'd either want or expect.
- Fixed the type of the WId parameter in the cmd line arg parsing to prevent
  it being clipped.

svn path=/trunk/kdebase/ksystraycmd/; revision=330471

+++ ksystraycmd/main.cpp
@@ -72,7 +72,7 @@ int main( int argc, char *argv[] )
       }
 
       bool ok=true;
-      int w = wid.toInt( &ok, base );
+      ulong w = wid.toULong( &ok, base );
       if ( ok )
 	  cmd.setTargetWindow( w );
       else {
